refact: 중복 로직 모듈화
